all: Fix implicit conversions from double to float
Either make the conversion explicit or use correct types, and enable
the compiler to emit warnings for it.

Warnings are not enabled for ports using libm (nrf, qemu-arm, samd)
becasue that just has too much warnings.  Also not enabled for
esp32 because external code like esp-idf/components/driver/mcpwm.c
has warnings, and not for esp8266 because it's compiler does not
support -Wfloat-conversion.

+++ ports/unix/modffi.c
@@ -378,7 +378,7 @@ STATIC mp_obj_t ffifunc_call(mp_obj_t self_in, size_t n_args, size_t n_kw, const
         #if MICROPY_PY_BUILTINS_FLOAT
         } else if (*argtype == 'f') {
             float *p = (float *)&values[i];
-            *p = mp_obj_get_float(a);
+            *p = (float) mp_obj_get_float(a);
         } else if (*argtype == 'd') {
             double *p = (double *)&values[i];
             *p = mp_obj_get_float(a);
